+++ npbloom-web/src/components/MainMenu.tsx
@@ -4,8 +4,9 @@ import { useOs } from '@mantine/hooks';
 import { IconDeviceFloppy, IconFileExport, IconFileImport, IconFolder, TablerIconsProps } from '@tabler/icons-react';
 import { MAIN_MENU_HEIGHT } from '../uiDimensions.ts';
 import useHotkeys from '@reecelucas/react-use-hotkeys';
-import { MarkCCommandedNodes, MarkCCommandingNodes, NodeSelectionInPlot, RemoveRelationMarkingsInSelectedTree,
-  RemoveAllRelationMarkings } from 'npbloom-core';
+import { FoldSelectedNodes, MarkCCommandedNodes, MarkCCommandingNodes, NodeSelectionInPlot,
+  RemoveRelationMarkingsInSelectedTree, RemoveAllRelationMarkings, UnfoldSelectedNodes, UnfoldSelectedNodesOneLevel,
+} from 'npbloom-core';
 import useUiState from '../useUiState.ts';
 import substituteOsAwareHotkey from './substituteOsAwareHotkey.ts';
 import Settings from './meta/Settings';
@@ -36,6 +37,8 @@ const MainMenu: React.FC = () => {
 
   const os = useOs();
 
+  const noNodesSelected = !(state.selection instanceof NodeSelectionInPlot);
+
   const oneNodeSelected = state.selection instanceof NodeSelectionInPlot &&
     state.selection.nodeIndicatorsAsArray.length === 1;
 
@@ -47,17 +50,29 @@ const MainMenu: React.FC = () => {
     if (oneNodeSelected) dispatch(new MarkCCommandedNodes());
   }
 
-  useHotkeys(['Control+o', 'Meta+o'], event => { event.preventDefault(); openFileLoadModal(); },
+  const foldSelectedNodes = () => dispatch(new FoldSelectedNodes());
+
+  const unfoldSelectedNodes = () => dispatch(new UnfoldSelectedNodes());
+
+  const unfoldSelectedNodesOneLevel = () => dispatch(new UnfoldSelectedNodesOneLevel());
+
+  useHotkeys(['Ctrl+O', 'Meta+O'], event => { event.preventDefault(); openFileLoadModal(); },
     { ignoredElementWhitelist: ['INPUT'] });
 
-  useHotkeys(['Control+s', 'Meta+s'], event => { event.preventDefault(); saveOrSaveAs(); },
+  useHotkeys(['Ctrl+S', 'Meta+S'], event => { event.preventDefault(); saveOrSaveAs(); },
+    { ignoredElementWhitelist: ['INPUT'] });
+
+  useHotkeys(['Ctrl+-', 'Meta+-'], event => { event.preventDefault(); foldSelectedNodes(); },
+    { ignoredElementWhitelist: ['INPUT'] });
+
+  useHotkeys(['Ctrl+=', 'Meta+='], event => { event.preventDefault(); unfoldSelectedNodes(); },
     { ignoredElementWhitelist: ['INPUT'] });
 
   const mainMenuElements: NamedMenuSection[] = [
     ['File', [
       [
-        { label: 'Open...', icon: IconFolder, hotkey: 'Ctrl-O', action: openFileLoadModal },
-        { label: activeFileName ? 'Save' : 'Save...', icon: IconDeviceFloppy, hotkey: 'Ctrl-S', action: saveOrSaveAs },
+        { label: 'Open...', icon: IconFolder, hotkey: 'Ctrl+O', action: openFileLoadModal },
+        { label: activeFileName ? 'Save' : 'Save...', icon: IconDeviceFloppy, hotkey: 'Ctrl+S', action: saveOrSaveAs },
         { label: 'Save as...', disabled: !activeFileName, action: openFileSaveModal },
       ],
       [
@@ -68,16 +83,19 @@ const MainMenu: React.FC = () => {
         { label: 'Export forest (experimental)...', icon: IconFileExport, action: openSystemFileSaveModal },
       ],
     ]],
-    ['Mark', [
+    ['View', [
       [
-        { label: 'C-commanding nodes', disabled: !oneNodeSelected, action: markCCommandingNodes },
-        { label: 'C-commanded nodes', disabled: !oneNodeSelected, action: markCCommandedNodes },
-      ],
-      [
+        { label: 'Mark C-commanding nodes', disabled: !oneNodeSelected, action: markCCommandingNodes },
+        { label: 'Mark C-commanded nodes', disabled: !oneNodeSelected, action: markCCommandedNodes },
         { label: 'Clear relation markings in selected tree',
           action: () => dispatch(new RemoveRelationMarkingsInSelectedTree()) },
         { label: 'Clear all relation markings',
           action: () => dispatch(new RemoveAllRelationMarkings()) },
+      ],
+      [
+        { label: 'Fold selected nodes', disabled: noNodesSelected, action: foldSelectedNodes, hotkey: 'Ctrl--' },
+        { label: 'Unfold selected nodes', disabled: noNodesSelected, action: unfoldSelectedNodes, hotkey: 'Ctrl-=' },
+        { label: 'Unfold selected nodes by one level', disabled: noNodesSelected, action: unfoldSelectedNodesOneLevel },
       ]
     ]],
   ]