+++ npbloom-web/src/components/meta/Settings.tsx
@@ -10,7 +10,7 @@ const Settings: React.FC = () => {
   const [opened, { open, close }] = useDisclosure(false);
   const { settingsState, settingsDispatch } = useContext(SettingsStateContext);
 
-  useHotkeys(['Control+,', 'Meta+,'], event => { event.preventDefault(); open(); },
+  useHotkeys(['Ctrl+,', 'Meta+,'], event => { event.preventDefault(); open(); },
     { ignoredElementWhitelist: ['INPUT'] });
 
   return <>